get_model_class_from_name
