--- custom_components/chihiros/chihiros_led_ch4_control/main/ch4.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-# Re-export the shared model so imports from either package succeed
-from .ch4 import Ch4
