+++ custom_components/chihiros/chihiros_led_control/device/LED/a2.py
@@ -1,6 +1,6 @@
 """A2 device Model."""
 
-from ..main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class AII(BaseDevice):
