+++ custom_components/chihiros/chihiros_led_control/device/LED/c2.py
@@ -1,6 +1,6 @@
 """CII device Model."""
 
-from ..main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class CII(BaseDevice):
