+++ custom_components/chihiros/chihiros_led_control/device/RGB/c2rgb.py
@@ -1,6 +1,6 @@
 """CII RGB device Model."""
 
-from ..main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class CIIRGB(BaseDevice):
