+++ custom_components/chihiros/chihiros_led_control/device/TINY/z_light_tiny.py
@@ -1,6 +1,6 @@
 """Z Light TINY device Model."""
 
-from ..main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class ZLightTiny(BaseDevice):
