+++ custom_components/chihiros/chihiros_led_control/device/WRGB/universal_wrgb.py
@@ -1,6 +1,6 @@
 """Universal WRGB device Model."""
 
-from ...chihiros_led_control.main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class UniversalWRGB(BaseDevice):
