+++ custom_components/chihiros/chihiros_led_control/device/WRGB/wrgb2_slim.py
@@ -1,6 +1,6 @@
 """WRGB II Slim device Model."""
 
-from ..main.base_device import BaseDevice
+from ...main.base_device import BaseDevice
 
 
 class WRGBIISlim(BaseDevice):
