+++ custom_components/chihiros/chihiros_led_control/device/commander4.py
@@ -1,6 +1,6 @@
 """Commander 4 device Model."""
 
-from ...chihiros_led_control.main.base_device import BaseDevice
+from ..main.base_device import BaseDevice
 
 
 class Commander4(BaseDevice):
